+++ src/components/Footer.js
@@ -1,0 +1,13 @@
+import React from "react";
+
+function Footer() {
+  return (
+    <div className="Footer-container">
+      <p className="Footer-text">
+        THE END
+      </p>
+    </div>
+  );
+}
+
+export default Footer;
