+++ src/components/Header.js
@@ -1,0 +1,12 @@
+import React from "react";
+import logo from "./logo.svg";
+
+function Header() {
+  return (
+    <div className="App-header">
+      <img src={logo} className="App-logo" alt="logo" />
+      <h2>Welcome to React</h2>
+    </div>
+  );
+}
+export default Header;
