+++ src/components/Intro.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+function Intro() {
+  return (
+    <p className="App-intro">
+      To get started, edit <code>src/App.js</code> and save to reload.
+    </p>
+  );
+}
+
+export default Intro;
