+++ src/components/Main.js
@@ -1,0 +1,16 @@
+import React from "react";
+import Section1 from "./Section1";
+import Section2 from "./Section1";
+import Section3 from "./Section1";
+
+function Main() {
+  return (
+    <div className="Main-container">
+      <Section1 />
+      <Section2 />
+      <Section3 />
+    </div>
+  );
+}
+
+export default Main;
