+++ src/components/Section1.js
@@ -1,0 +1,16 @@
+import React from "react";
+
+function Section1() {
+  return (
+    <div className="Section-content">
+      <p>
+        Lorem ipsum dolor sit amet, consectetur adipiscing elit. Suspendisse v
+        estibulum nisi nec est iaculis, ut tristique arcu hendrerit.
+        Cras velit nibh, mollis sit amet odio ut, dapibus convallis nisi.
+        Donec id egestas ipsum. Duis id egestas justo, ac porta diam.
+      </p>
+    </div>
+  );
+}
+
+export default Section1;
