+++ src/components/Section2.js
@@ -1,0 +1,15 @@
+import React from "react";
+
+function Section2() {
+  return (
+    <div className="Section-content">
+      <p>
+        Suspendisse maximus, odio vel tincidunt viverra, sem dui feugiat mauris,
+        eu ultricies lorem dui vitae est. Proin ac vehicula velit.
+        Vestibulum eget diam ullamcorper, consectetur est at, imperdiet turpis.
+      </p>
+    </div>
+  );
+}
+
+export default Section2;
