+++ src/components/Section3.js
@@ -1,0 +1,18 @@
+import React from "react";
+
+function Section3() {
+  return (
+    <div className="Section-content">
+      <p>
+      Duis dictum fringilla nulla et tincidunt. Praesent rutrum ac nisi ut tempor.
+      Curabitur non velit eget arcu semper suscipit.
+      Vestibulum scelerisque laoreet metus quis dapibus.
+      Proin ultricies velit eget nisl elementum sollicitudin sed eget tortor.
+      Vivamus aliquam placerat ante sed posuere. Sed finibus vestibulum augue id interdum.
+      Suspendisse potenti.
+      </p>
+    </div>
+  );
+}
+
+export default Section3;
